server/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
mongoengine==0.28.2
bcrypt==4.1.2
PyJWT==2.8.0
torch==2.2.2
llama-index==0.10.17
transformers==4.38.2
certifi==2023.11.17